Flask inverts image and returns it
Flask is able to process the image, and will invert the colours as a proof of concept. Images saved in the 'Images' folder currently, but may need a better solution for image management.

Inverting will obviously be replaced, but has been used to test that I can save and modify images and the correct one is returned to React